move server app to top level so Spring finds JPA repository - this qualifies as a hacked-up reaction to my failure to understand how to config
