+++ ctk-server/src/test/java/org/ga4gh/ctk/server/CtkServerApplicationTests.java
@@ -1,5 +1,6 @@
 package org.ga4gh.ctk.server;
 
+import org.ga4gh.ctk.*;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.test.context.web.WebAppConfiguration;
